fix(report): corrige: alteração no filtro refletindo no report renderizado
